refactor(search): move search-results-page to views directory
