+++ src/components/filters/components/filters-date-slider/hooks/useDateRangeContext.tsx
@@ -1,6 +1,6 @@
 import React, { useEffect, useMemo, useState } from 'react';
 import { addMissingYears } from '../helpers';
-import { FilterItem } from 'src/components/search-results-page/components/filters/types';
+import { FilterItem } from 'src/views/search-results-page/components/filters/types';
 
 export interface ContextProps {
   colorScheme: string;
