+++ src/views/search-results-page/components/card/helpers.ts
