+++ src/views/search-results-page/components/card/index.tsx
