+++ src/views/search-results-page/components/card/metadata-accordion/index.tsx
