+++ src/views/search-results-page/components/card/source-logo.tsx
