+++ src/views/search-results-page/components/card/styles.ts
