+++ src/views/search-results-page/components/card/topic-categories/index.tsx
