+++ src/views/search-results-page/components/count/index.tsx
