+++ src/views/search-results-page/components/filters/config.ts
