+++ src/views/search-results-page/components/filters/helpers.ts
