+++ src/views/search-results-page/components/filters/types.ts
