+++ src/views/search-results-page/components/filters/utils/queries.ts
