+++ src/views/search-results-page/components/metadata-score-toggle/index.tsx
