+++ src/views/search-results-page/components/pagination/index.tsx
