+++ src/views/search-results-page/helpers.ts
